pyproject.toml
[project]
name = "geodjango-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "gdal>=3.11.4",
    "python-dotenv>=1.1.1",
    "spatialite>=0.0.3",
]
